enhance: Move injectables to new data module
