+++ src/cleantest/data/pkg/charmlib.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python3
-# Copyright 2022 Jason C. Nucciarone, Canonical Ltd.
+# Copyright 2023 Jason C. Nucciarone, Canonical Ltd.
 # See LICENSE file for licensing details.
 
 """Manager for installing charm libraries inside remote processes."""
@@ -13,15 +13,12 @@ from shutil import which
 from typing import List, Union
 
 from cleantest.meta import BasePackage, BasePackageError, InjectableData
-from cleantest.pkg.handler import snap
-
-from ._mixins import SnapdSupport
+from cleantest.meta.mixins import SnapdSupport
+from cleantest.utils import snap
 
 
 class CharmlibPackageError(BasePackageError):
     """Base error for Charmlib package handler."""
-
-    ...
 
 
 class Charmlib(BasePackage, SnapdSupport):
