--- src/cleantest/hooks/__init__.py
@@ -1,8 +1,0 @@
-#!/usr/bin/env python3
-# Copyright 2022 Canonical Ltd.
-# See LICENSE file for licensing details.
-
-from .startenv import StartEnvHook
-from .starttestlet import StartTestletHook
-from .stopenv import StopEnvHook
-from .stoptestlet import StopTestletHook
