--- src/cleantest/hooks/data/__init__.py
@@ -1,6 +1,0 @@
-#!/usr/bin/env python3
-# Copyright 2022 Jason C. Nucciarone, Canonical Ltd.
-# See LICENSE file for licensing details.
-
-from .directory import Dir
-from .file import File
