+++ src/cleantest/meta/mixins/__init__.py
@@ -3,3 +3,4 @@
 # See LICENSE file for licensing details.
 
 from .dict_ops import DictOps
+from .snapd_support import SnapdSupport
