--- src/cleantest/pkg/_mixins/__init__.py
@@ -1,5 +1,0 @@
-#!/usr/bin/env python3
-# Copyright 2022 Jason C. Nucciarone, Canonical Ltd.
-# See LICENSE file for licensing details.
-
-from .snapd_support import SnapdSupport
