--- src/cleantest/pkg/handler/__init__.py
@@ -1,5 +1,0 @@
-#!/usr/bin/env python3
-# Copyright 2022 Jason C. Nucciarone, Canonical Ltd.
-# See LICENSE file for licensing details.
-
-from . import snap_handler as snap
